+++ ChangeLog.rst
@@ -9,6 +9,8 @@ Unreleased
   releasing next version, the "Unreleased" header will be replaced with
   appropriate version header and this help text will be removed.
 
+- Remove AddressManager
+- Split address into mutable and immutable address
 
 Version 2.0.0
 -------------